fix(events): suppress tab coalescing after a native-space switch
Closes the architect re-review's residual: a targeted reconcile
(focus-changed / app-activation) racing the bulk reconcileAll during a
native-space switch kept coalesceTabs true, so a tab carrier spanning
two spaces at the identical tiled frame could false-merge. Now every
reconcile within a short grace window after a native-space change
suppresses coalescing (reconcileAll already did).

- EventLoop.lastNativeSpaceChange stamped in nativeSpaceChanged;
  reconcile ANDs in `!recentSpaceSwitch` (0.75 s grace). A genuine
  tab switch inside the window self-heals to destroy + create.
- Surface the two accepted same-frame false-merge edges (OverlapStack
  pile + cross-space carrier) in the design-decisions row.
- Tighten TabReconciler's bias wording to the either-side reality.
- Docs: user-guide native-tabs note (Ignore Rules), AGENTS.md §5
  guardrail (one NSWindow per tab, only active tab AX-visible,
  temporal coalesce).
- Extract reconcile() into EventLoop+Reconcile.swift (the grace check
  pushed EventLoop.swift past the 350-line ceiling).
