+++ docs/design-decisions.md
@@ -56,7 +56,7 @@ planned escape hatch; it is not a wontfix dumping ground.
 | An app with **no bundle identifier** (a rare unbundled or legacy helper process) can't be targeted by an app rule (`float_rules`, `ignore_rules`, `app_rules`) or by `pull_or_spawn`; its windows match no rule and tile normally. | App rules key on the bundle identifier because it is the only stable, locale- and rename-proof identity — the display name is presentation, not identity. A name-based fallback for id-less apps would reintroduce exactly the ambiguity (two apps sharing a display name, locale drift) that keying on the id removes, and split the vocabulary into two identity schemes. | Window identity is `NSRunningApplication.bundleIdentifier`, captured once at attach and stored on `ManagedWindow.appBundleID` (`AppRef`); a nil id stays nil and matches no rule ([#262](https://github.com/hajiboy95/KiwiDesk/issues/262)). | None — effectively every user-facing app is bundled and has an identifier; the blind spot is a handful of window-less helper processes not worth managing. |
 | A **thick** focus border at **small gaps** makes neighbouring windows' rings touch or visually merge, and at 0 pt gaps a ring paints over the adjacent neighbour's edge. | The ring is a pure post-layout overlay with no gap coupling (coupling would leak a visual setting into pure layout math), and it is drawn **pure outset** — entirely outside the window frame, so its outward reach is the full `width` per side and rings meet when `2·width ≥ gap`. The worst case paints over a *neighbour's* edge (reads as a halo), never the focused window's own content — outset never touches it. An automatic minimum-gap floor was rejected: silently widening a user's `gap = 0` the moment borders turn on is paternalistic and changes a value the user set; `border.fit_gaps` stays the opt-in escape hatch. | The border is rendered independently of layout ([#303](https://github.com/hajiboy95/KiwiDesk/issues/303)); the width is bounded to 20 pt, but gaps and border width are set separately. | Set gaps at least as wide as the border (or run `border.fit_gaps`), or lower the width; the default 2 pt border is imperceptible at any typical gap. |
 | A focused-window shortcut (`focus`, `swap`, `resize`, `make_floating`, `move_to_space`, `move_to_track`, `stack.promote`/`demote`, …) is **rejected without acting** while an ignored panel (Ghostty's quick terminal) or an unmanaged/ignored app holds the foreground — and, for the brief moment during an app-activation or self-raise race before the OS frontmost app catches up to KiwiDesk's focused window. | Acting on the implicit focused window while a *different* window is frontmost would silently mutate a window the user cannot see. Failing closed until the OS foreground genuinely matches the managed focused window is the safe answer: a transient rejected shortcut during activation is far less harmful than a hidden mutation. An unmanaged panel that never emits any focus notification cannot set the ignored-panel latch, so a shortcut fired while it is up relies on the frontmost-pid check alone — the deliberately preserved limit from the Ghostty panel work. | One semantic preflight at `KiwiCore.execute` (`FocusedCommandPolicy` classifies the focused commands; `focusedCommandDenial` requires the OS frontmost pid to equal the focused managed window's pid, the event loop to still observe it, and no ignored-panel latch for it) — shared by Lua, CLI, and IPC, else `no managed window is currently focused` ([#292](https://github.com/hajiboy95/KiwiDesk/issues/292)). | Bring the managed window back to the foreground (click it, or dismiss the panel so its app re-reports a managed window) and re-issue. Global config, `focus_space`, spawns, profile ops, and explicit App Bar clicks are never gated. |
-| A **native-tab window** (Finder, Terminal, Ghostty) is managed as **one** window that follows whichever tab is active — its tabs can't be split into separate tiles, and there is one App Bar item for the group, not one per tab. | Native tabs are separate `NSWindow`s the app owns, with only the active tab ever visible to AX; splitting a tab into its own tile would need cross-process `NSWindow` reparenting KiwiDesk cannot perform. Managing the group as one slot that re-keys to the active tab is the honest model — one tile, one App Bar item, and no spurious tile or focus jump on switch/close. Expanding tabs as App Bar sub-items was rejected: the app's own tab bar already does that, and it would be un-Mac-like. | Background tabs never appear in `kAXWindowsAttribute` and mint a fresh `CGWindowID` per switch ([#308](https://github.com/hajiboy95/KiwiDesk/issues/308) probe); the event loop coalesces the same-frame vanish/appear into `.windowRekeyed` (`TabReconciler`), preserving the flat one-slot-per-group state. | None in-app — a cross-process split is impossible. Whole-app opt-out via `ignore_rules` if a specific app's tab behavior misbehaves. |
+| A **native-tab window** (Finder, Terminal, Ghostty) is managed as **one** window that follows whichever tab is active — its tabs can't be split into separate tiles, and there is one App Bar item for the group, not one per tab. | Native tabs are separate `NSWindow`s the app owns, with only the active tab ever visible to AX; splitting a tab into its own tile would need cross-process `NSWindow` reparenting KiwiDesk cannot perform. Managing the group as one slot that re-keys to the active tab is the honest model — one tile, one App Bar item, and no spurious tile or focus jump on switch/close. Expanding tabs as App Bar sub-items was rejected: the app's own tab bar already does that, and it would be un-Mac-like. | Background tabs never appear in `kAXWindowsAttribute` and mint a fresh `CGWindowID` per switch ([#308](https://github.com/hajiboy95/KiwiDesk/issues/308) probe); the event loop coalesces the same-frame vanish/appear into `.windowRekeyed` (`TabReconciler`), preserving the flat one-slot-per-group state. Detection is temporal (a tab group on either side + same frame within tolerance, in one reconcile pass), so two narrow same-frame false-merge edges are accepted: same-app windows deliberately stacked in an `OverlapStack` pile, and a tab carrier spanning two native Spaces at the identical tiled frame (suppressed by a post-space-switch grace window). Both need the vanish and appear in one pass and self-heal on the next reconcile. | None in-app — a cross-process split is impossible. Whole-app opt-out via `ignore_rules` if a specific app's tab behavior misbehaves. |
 
 ### Blocked by macOS (SIP)
